Revert "Reduce the size of viswall_t by 8 bytes"
This reverts commit ec869a407e4bd35d8258f61b578f9a2ff570ac7a.

+++ r_phase9.c
@@ -110,7 +110,7 @@ static void R_UpdateCache(void)
             }
         }
 
-        if ((int8_t)wall->ceilingpicnum == -1) {
+        if (wall->ceilingpicnum == -1) {
             continue;
         }
